mcp/requirements.txt
mcp>=0.9.0
starlette>=0.37.0
uvicorn>=0.24.0
sse-starlette>=1.8.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0
